Reorganizing api functions and models
