+++ front_end/src/components/auth/DFRoutes.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from "react";
 import { Navigate, Route, Routes } from "react-router-dom";
-import { Permission } from "../../sdk";
+import { Permission } from "../../api/models";
 import { useAuth } from "../../utils/AuthContext";
 import { hasPermission } from "../../utils/helperFunctions";
 import { PrivateRoute } from "../../utils/PrivateRoute";
