--- front_end/src/sdk.ts
@@ -1,93 +1,0 @@
-import { makePostRequest, makeGetRequest } from "./utils/apiUtils";
-
-export enum Permission {
-  NONE = "0",
-  CUSTOMER = "1",
-  WORKER = "2",
-  MANAGER = "3",
-  ADMIN = "4",
-}
-export enum IngredientType {
-  MILK = "MILK",
-  FLAVOR = "FLAVOR",
-  TOPPING = "TOPPING",
-  SWEETENER = "SWEETENER",
-  ADDIN = "ADDIN",
-  COFFEE = "COFFEE",
-  SHOT = "SHOT",
-}
-export interface User {
-  email: string;
-  firstName: string;
-  lastName: string;
-  permissions: Permission;
-  userId: string;
-}
-export interface Ingredient {
-  IngredientId: string;
-  Kind: IngredientType;
-  Name: string;
-  Price: number | null | undefined;
-  Stock: number | null | undefined;
-  Upcharge: number;
-}
-
-export interface Drink {
-  menuItem: MenuItem;
-  recipe: RecipeItem[];
-}
-
-export interface RecipeItem {
-  ingredient: Ingredient;
-  quantity: number;
-}
-
-export interface MappingOfIngredientToQuantity {
-  [ingredientId: string]: number;
-}
-
-export interface MenuItem {
-  MenuId: number;
-  Name: string;
-  Recipe: MappingOfIngredientToQuantity;
-  Price: number;
-  Active: boolean;
-  ImagePath: string;
-}
-
-export interface LoginResponse {
-  token: string;
-  error: string;
-  user: User;
-}
-
-export function loginUser(
-  password: string,
-  email: string
-): Promise<LoginResponse> {
-  return makePostRequest("auth/token", { password, email });
-}
-
-export function getIngredients(): Promise<Ingredient[]> {
-  return makeGetRequest("ingredients/").then((res) => res.ingredients);
-}
-
-export function getMenuItems(): Promise<MenuItem[]> {
-  return makeGetRequest("menuitems/").then((res) => res.menuitems);
-}
-
-export function getIngredientKinds(): Promise<IngredientType[]> {
-  return makeGetRequest("ingredients/kind").then((res) => res.kinds);
-}
-
-export function getIngredientById(id: string): Promise<Ingredient> {
-  return makeGetRequest(`ingredients/${id}`).then((res) => res.ingredient);
-}
-
-export function createIngredient(
-  ingredient: CreateIngredientType
-): Promise<Ingredient> {
-  return makePostRequest("ingredients/create", ingredient);
-}
-
-export type CreateIngredientType = Omit<Ingredient, "IngredientId">;
