+++ front_end/src/utils/apiUtils.ts
@@ -16,7 +16,7 @@ export function makeGetRequest(path: string): Promise<any> {
     method: "GET",
     headers: {
       "Content-Type": "application/json",
-      Authorization: "Bearer " + localStorage.getItem("token"),
+      Authorization: `${localStorage.getItem("token")}`
     },
   }).then((res) => res.json());
 }
