+++ front_end/src/utils/helperFunctions.ts
@@ -1,7 +1,13 @@
 import { TextFieldProps } from "@mui/material";
 import { FormikValues, useFormik } from "formik";
 import { matchPath, useLocation } from "react-router";
-import { MenuItem, RecipeItem, Ingredient, Permission, Drink } from "../sdk";
+import {
+  Ingredient,
+  MenuItem,
+  Permission,
+  RecipeItem,
+  Drink,
+} from "../api/models";
 
 export function useRouteMatch(patterns: string[]) {
   const { pathname } = useLocation();
